116 исправить баги с устновкой pageid (#117)
* #116 change pageId type from int to long

* #116 null option

* #116 validation of the case when pageId < 0

* #116 fix greterThanOrEqualTo

+++ S4C.Services/Services/GetGamesDataService/HttpRequestMethodDictionaries/PartnerInterfaceHttpRequestsMethodDictionary.cs
@@ -18,7 +18,7 @@ namespace C4S.Services.Services.GetGamesDataService.HttpRequestMethodDictionarie
         /// </returns>
         public static HttpRequestMessage GetAppReport(
             string authorizationToken,
-            int pageId,
+            long pageId,
             DateTimeRange period)
         {
             var dateTimeFormat = "yyyy-MM-dd";
